requirements.txt
flask
httpx
gunicorn
sqlalchemy
psycopg2-binary
redis
httpx
pytest
python-dotenv
